Made all sprites use colorkeys instead of alpha.
SDL (and therefore pygame) are not very efficient when it comes to blitting
surfaces that contain an alpha channel (transparent or partially transparent
pixels).  They're much faster when one colour in an image represents
transparent areas.  This game apparently runs slowly on the Pandora, so this
will hopefully make it work more smoothly.